Fix antonym quizzes of concepts with multiple grammatical forms.
When quizzing the antonym of one grammatical form of a concept with multiple grammatical forms, Toisto would accept all grammatical forms of the antonym concept. For example, when quizzing the antonym of halvempi in Finnish (cheaper), Toisto would accept kallis (expensive), kalliimpi (more expensive), and kallein (most expensive) as answer. Fixed to only accept the same grammatical form of the antonym concept, kalliimpi (more expensive) in the example.

Fixes #893.